// src/components/ui/questionnaire/types.ts
type QuestionnaireChoiceDefinition = {
  disabled?: boolean
  value: string
}

type QuestionnaireItemDefinition = {
  choices?: readonly QuestionnaireChoiceDefinition[]
  disabled?: boolean
  name: string
  required?: boolean
}